fix spacing, log in debug only, abstract optional logging component
