+++ Samples/Server/Unity-FractalStream/Assets/Scripts/WebRTCServerDebug.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: de57184637e8d454386d0d895f2875d4
+timeCreated: 1506630927
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
